Moved pending to userlist
