+++ client/webbchat/src/components/userList/pending/App.js
@@ -1,7 +1,7 @@
 
 import { bindActionCreators } from 'redux';
 import { connect } from 'react-redux';
-import * as actionsCreators from '../../actions/actionCreators';
+import * as actionsCreators from '../../../actions/actionCreators';
 import Users from './Users';
 
 const mapStateToProps = (state) => {
